Added promises in under 100 lines
This promise implementation is far from correct. But the goal behind these are not to be correct. They are meant to teach and show what is possible in a small amount of code. In this case I think the shortcuts are acceptable. I do actually think a correct (though not complete) implementation is possible in under 100 lines. It might actually be fun to revisit the promises later and fix them to be correct.